change signature to return the step and move up
