+++ api-examples/transformer_utils.py
@@ -15,14 +15,6 @@ import glob
 import json
 
 
-def find_latest_checkpoint(checkpoint_dir: str, wildcard="checkpoint") -> str:
-    step_num = 0
-    for f in glob.glob(os.path.join(checkpoint_dir, f"{wildcard}*")):
-        this_step_num = int(f.split("-")[-1])
-        if this_step_num > step_num:
-            checkpoint = f
-            step_num = this_step_num
-    return checkpoint
 
 
 class TripletLoss(nn.Module):
